+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Improvements:
 * [OLMIS-6005](https://openlmis.atlassian.net/browse/OLMIS-6005): Improved performance of GET /api/supervisoryNodes/{supervisoryNode_id}/supervisingUsers endpoint.
 * [OLMIS-5853](https://openlmis.atlassian.net/browse/OLMIS-5853): Add Unique Name constraint on Supervisory Node
 * [OLMIS-5852](https://openlmis.atlassian.net/browse/OLMIS-5852): Add Case Insensitive Unique Code constraint on Supervisory Node
+* [OLMIS-6004](https://openlmis.atlassian.net/browse/OLMIS-6004) : Add sortable parameter to program search api
 12.0.0 / 2018-12-12
 ===================